Make polaris-spinner respect the color property
