+++ addon/templates/components/polaris-icon.hbs
@@ -6,5 +6,6 @@
     class="Polaris-Icon__Svg"
     focusable="false"
     aria-hidden="true"
+    id=svgElementId
   }}
 {{/if}}
