+++ addon/templates/components/polaris-spinner.hbs
@@ -1,5 +1,6 @@
 {{svg-jar spinnerSource
   role="status"
+  id=svgElementId
   class=spinnerClass
   aria-label=accessibilityLabel
 }}
